move is_rpm_based_os to unit_tests module
